SmarterCoffee/extras/requirements.txt
progressbar2
console-menu
pyaml
